Extract Reducer from AppMessage class
Was walking the dog the other night and realized it's a little weird
having a mostly unrelated reducer slapped into the AppMessage class.
Originally my though had been to consolidate all message logic into that
one class but, really, that's what modules are for.

So I've pulled the reducer into the context file and moved the tests
over there. I hope that will be my last movement on the issue.
