+++ src/client/classes/AppMessage.ts
@@ -1,5 +1,3 @@
-import { Reducer } from 'react';
-
 /**
  * Possible types of message to display to the user
  */
@@ -18,25 +16,6 @@ export enum MessageAction {
   push = 'push',
   clear = 'clear',
 }
-
-/**
- * Defines the state of the message reducer
- */
-
-export interface MessageReducerState {
-  queue: AppMessage[];
-  currentMessage: AppMessage;
-}
-
-/**
- * Defines that kinds of actions that can be accepted by the reducer
- */
-
-export interface MessageReducerAction {
-  type: MessageAction;
-  message?: AppMessage;
-}
-
 /**
  * Used for displaying messages on the client
  */
@@ -62,40 +41,4 @@ export class AppMessage {
     this.variant = variant;
     this.text = message;
   }
-
-  /**
-   * handles queueing logic for the top-level app component
-   */
-
-  public static reducer: Reducer<MessageReducerState, MessageReducerAction> = (
-    state,
-    action
-  ): MessageReducerState => {
-    const { currentMessage, queue } = state;
-    switch (action.type) {
-      case (MessageAction.push): {
-        if (!currentMessage) {
-          return ({
-            ...state,
-            currentMessage: action.message,
-          });
-        }
-        const newQueue = [...queue, action.message];
-        return {
-          ...state,
-          queue: newQueue,
-        };
-      }
-      case (MessageAction.clear): {
-        const nextQueue = [...queue];
-        const nextMessage = nextQueue.shift();
-        return {
-          queue: nextQueue,
-          currentMessage: nextMessage,
-        };
-      }
-      default:
-        return state;
-    }
-  };
 }
